move useSubmitSongFromRemoteFile()
