+++ projects/graphql-client/src/index.ts
@@ -45,3 +45,4 @@ export * from "./remote/mutations/useIssueAuthToken"
 export * from "./remote/mutations/useRestorePassword"
 export * from "./remote/mutations/useSongFileDuplicates"
 export * from "./remote/mutations/useGenerateUploadableUrl"
+export * from "./remote/mutations/useSubmitSongFromRemoteFile"
